Fix PR suggestions
