+++ cli/ballerina-cli/src/main/java/io/ballerina/cli/task/CompileTask.java
@@ -85,7 +85,7 @@ public class CompileTask implements Task {
 
             PackageResolution packageResolution = project.currentPackage().getResolution();
 
-            if(project.currentPackage().compilationOptions().dumpRawGraphs()) {
+            if (project.currentPackage().compilationOptions().dumpRawGraphs()) {
                 packageResolution.dumpGraphs(out);
             }
 
@@ -109,6 +109,8 @@ public class CompileTask implements Task {
                 }
             }
 
+            // We dump the raw graphs twice only if code generator/modifier plugins are engaged
+            // since the package has changed now
             if (packageResolution != project.currentPackage().getResolution()) {
                 packageResolution = project.currentPackage().getResolution();
                 if (project.currentPackage().compilationOptions().dumpRawGraphs()) {
